requirements.txt
flask
twilio
openai
requests
beautifulsoup4
gunicorn
python-dotenv